+++ src/axios/audio_classification.tsx
@@ -1,25 +1,20 @@
-/* 도로 위성 사진 분류 - 위성 분류 */
+/* 음성단어분류 - 음성 분류 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import base64DataToFile from './base64DataToFile'
+import { Keyword } from '@/components';
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 
-const satelliteClassification = () => {
+const audioClassification = () => {
   let data = {
-    file: base64DataToFile(detailData, '사진이름', 'image/png'),
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/f9299/inference',
+    file: base64DataToFile(detailData, '오디오이름', 'audio/wav'),
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/c68e0/inference',
   }
+
   setLoading(true)
-  let class_info: any = {
-    freeway: '고속도로',
-    intersection: '교차로',
-    overpass: '고가도로',
-    parkinglot: '주차장',
-    runway: '일반도로',
-  }
 
   axios
     .post('/inference/file_req_ajx', data, {
@@ -35,7 +30,10 @@ const satelliteClassification = () => {
         if (response_data == null) {
           response_data = json.response.inference
         }
-        response_data = class_info[response_data]
+        // 결과 들어가는 부분
+        // $(".result_alert").html(response_data);
+        // $("div.wrap_next").addClass("show_alert_pass");
+        <Keyword option={1} label={response_data}/>
       } else {
         alert('API 호출에 실패했습니다.')
       }
@@ -48,4 +46,4 @@ const satelliteClassification = () => {
     })
 }
 
-export default satelliteClassification
+export default audioClassification
