+++ src/axios/audio_clustering.tsx
@@ -3,6 +3,7 @@ import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import base64DataToFile from './base64DataToFile'
+import { Keyword } from '@/components';
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
@@ -36,9 +37,10 @@ const audioClustering = () => {
           response_data = json.response.inference
         }
         /* 결과 */
-        response_data = audio_info[response_data]
+        response_data = audio_info[response_data];
         // $(".result_alert").html(response_data);
         // $("div.wrap_next").addClass("show_alert_pass");
+        <Keyword option={1} label={response_data} />
       } else {
         alert('API 호출에 실패했습니다.')
       }
