+++ src/axios/binary_anomaly.tsx
