+++ src/axios/binary_regression.tsx
@@ -1,17 +1,21 @@
-/* 픽셀의 다중 스펙트럼 값을 이용한 이상탐지 - 위성 이상탐지 */
+/* 악성코드 감염 예측 - 바이너리 회귀 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
+import { Keyword } from '@/components'
+// import base64DataToFile from './base64DataToFile'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 
-const satelliteAnomaly = () => {
+const binaryRegression = () => {
   let data = JSON.stringify({
-    word: detailData.join(),
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/69e0b/inference',
+    word: detailData,
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/a142d/inference',
   })
+
   setLoading(true)
+
   axios
     .post('/inference/log_req_ajx', data, {
       headers: {
@@ -24,16 +28,8 @@ const satelliteAnomaly = () => {
       if (json.res == 'true') {
         let response_data = json.response.data
         if (response_data == null) {
-          response_data = json.response.inference
-        }
-
-        /* 결과 */
-        if (response_data == 'normal data') {
-          // 정상 결과 들어가는 부분
-          // $("div.wrap_next").addClass("show_alert_pass");
-        } else if (response_data == 'abnormal data') {
-          // 파손 결과 들어가는 부분
-          // $("div.wrap_next").addClass("show_alert_nonpass");
+          response_data = json.response.inference;
+          <Keyword option={1} label={response_data} />
         } else {
           alert('API 호출에 실패했습니다.')
         }
@@ -47,4 +43,4 @@ const satelliteAnomaly = () => {
     })
 }
 
-export default satelliteAnomaly
+export default binaryRegression
