+++ src/axios/image_anomaly.tsx
@@ -1,25 +1,19 @@
-/* 만화 얼굴 군집화 - 이미지 군집화 */
+/* 노후 시설물 이미지를 이용한 이상탐지 - 이미지 이상탐지 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom, resultAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import base64DataToFile from './base64DataToFile'
+import { Keyword } from '@/components';
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 const setResult = useSetRecoilState(resultAtom)
 
-const imageCluster = () => {
+const imageAnomaly = () => {
   // 이미지 이름이랑, 밈타입 알아야할듯?
   let data = {
     file: base64DataToFile(detailData, '이미지이름', 'image/jpeg'),
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/48f97/inference',
-  }
-
-  let class_info: any = {
-    // 띄어쓰기 하려고 문자형으로 함
-    'perm hair': '웨이브 머리',
-    'straight hair': '직모',
-    sunglasses: '선글라스',
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/0897f/inference',
   }
 
   setLoading(true)
@@ -33,15 +27,22 @@ const imageCluster = () => {
       responseType: 'json',
     })
     .then(res => {
+      // $("#input_log").val(JSON.stringify(json.request));
       let json = res.data
       if (json.res == 'true') {
+        // $('#output_log').val(JSON.stringify(json.response))
+        // $('div.wrap_logText').addClass('show_log')
         var response_data = json.response.data
         if (response_data == null) {
           response_data = json.response.inference
         }
-        response_data = class_info[response_data]
-        // $('.result_alert').html(response_data)
-        // $('div.wrap_next').addClass('show_alert_pass')
+        if (response_data == 'original') {
+          // $('div.wrap_next').addClass('show_alert_pass')
+          <Keyword option={1} label={'정상 블록'} />
+        } else if (response_data == 'discard') {
+          // $('div.wrap_next').addClass('show_alert_nonpass')
+          <Keyword option={2} label={'파손 블록'} />
+        }
       } else {
         alert('API 호출에 실패했습니다.')
       }
@@ -54,4 +55,4 @@ const imageCluster = () => {
     })
 }
 
-export default imageCluster
+export default imageAnomaly
