+++ src/axios/image_classification.tsx
