+++ src/axios/log_anomaly.tsx
@@ -1,18 +1,17 @@
-/* 악성코드 감염 예측 - 바이너리 회귀 */
+/* 신용카드 사기 탐지 - log 이상탐지 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
-// import base64DataToFile from './base64DataToFile'
+import { Keyword } from '@/components'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 
-const binaryRegression = () => {
+const logAnomaly = () => {
   let data = JSON.stringify({
-    word: detailData,
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/a142d/inference',
+    word: detailData.join(),
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/d0fb3/inference',
   })
-
   setLoading(true)
 
   axios
@@ -25,9 +24,22 @@ const binaryRegression = () => {
     .then(res => {
       let json = res.data
       if (json.res == 'true') {
+        // $("#output_log").val(JSON.stringify(json.response));
+        // $("div.wrap_logText").addClass("show_log");
         let response_data = json.response.data
         if (response_data == null) {
           response_data = json.response.inference
+        }
+
+        /* 결과 */
+        if (response_data == 'normal_transaction') {
+          // 정상 결과 들어가는 부분
+          // $("div.wrap_next").addClass("show_alert_pass");
+          <Keyword option={1} label={'정상거래'} />
+        } else if (response_data == 'fraudulent_transaction') {
+          // 파손 결과 들어가는 부분
+          // $("div.wrap_next").addClass("show_alert_nonpass");
+          <Keyword option={2} label={'부정거래'} />
         } else {
           alert('API 호출에 실패했습니다.')
         }
@@ -41,4 +53,4 @@ const binaryRegression = () => {
     })
 }
 
-export default binaryRegression
+export default logAnomaly
