+++ src/axios/log_clustering.tsx
@@ -2,6 +2,7 @@
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
+import { Keyword } from '@/components';
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
@@ -34,7 +35,8 @@ const logClustering = () => {
         if (response_data == null) {
           response_data = json.response.inference
         }
-        response_data = class_info[response_data]
+        response_data = class_info[response_data];
+        <Keyword option={1} label={response_data} />
         // 결과 들어가는 부분
         //  $(".result_alert").html(response_data);
         // $("div.wrap_next").addClass("show_alert_pass");
