+++ src/axios/log_regression.tsx
@@ -1,24 +1,23 @@
-/* 음성단어분류 - 음성 분류 */
+/* NBA 선수 연봉 예측 - log 회귀 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
-import base64DataToFile from './base64DataToFile'
+import { Keyword } from '@/components'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 
-const audioClassification = () => {
-  let data = {
-    file: base64DataToFile(detailData, '오디오이름', 'audio/wav'),
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/c68e0/inference',
-  }
-
+const logRegression = () => {
+  let data = JSON.stringify({
+    word: detailData.join(),
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/6dbca/inference',
+  })
   setLoading(true)
 
   axios
-    .post('/inference/file_req_ajx', data, {
+    .post('/inference/log_req_ajx', data, {
       headers: {
-        'Content-Type': 'multipart/form-data',
+        'Content-Type': 'application/json',
       },
       responseType: 'json',
     })
@@ -30,8 +29,9 @@ const audioClassification = () => {
           response_data = json.response.inference
         }
         // 결과 들어가는 부분
-        // $(".result_alert").html(response_data);
+        //  $(".result_alert").html(response_data);
         // $("div.wrap_next").addClass("show_alert_pass");
+        <Keyword option={1} label={response_data} />
       } else {
         alert('API 호출에 실패했습니다.')
       }
@@ -44,4 +44,4 @@ const audioClassification = () => {
     })
 }
 
-export default audioClassification
+export default logRegression
