+++ src/axios/satellite_clustering.tsx
@@ -3,6 +3,7 @@ import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import base64DataToFile from './base64DataToFile'
+import { Keyword } from '@/components'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
@@ -29,7 +30,8 @@ const satelliteClustering = () => {
         if (response_data == null) {
           response_data = json.response.inference
         }
-        response_data = class_info[response_data]
+        response_data = class_info[response_data];
+        <Keyword option={1} label={response_data} />
       } else {
         alert('API 호출에 실패했습니다.')
       }
