+++ src/axios/satellite_regression.tsx
@@ -3,6 +3,7 @@ import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import base64DataToFile from './base64DataToFile'
+import { Keyword } from '@/components'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
@@ -27,6 +28,7 @@ const satelliteRegression = () => {
         if (response_data == null) {
           response_data = json.response.inference
         }
+        <Keyword option={1} label={response_data} />
       } else {
         alert('API 호출에 실패했습니다.')
       }
